pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openapi-locustgen"
version = "0.1.0"
description = "Generate thin API clients and Locust load test scaffolding from OpenAPI specs"
authors = [{ name = "OpenAI Assistant" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
openapi-locustgen = "openapi_locustgen.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["openapi_locustgen"]
